Rename module 'request' -> 'http'
